Added handler for process info, continued initial work on interface.
Added a handler to get the static process info. This is called by the
client upon receiving the first status for a process.
Continued some initial work on making the UI useful.
Fixed the repeating HTTP get for status to not spin as fast as it could
forever on a failure.